implemented csv loader 2.0

+++ hg-core/src/main/java/ua/vlad/hg/core/dto/LatLng.java
@@ -3,11 +3,13 @@ package ua.vlad.hg.core.dto;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 public class LatLng {
 
-    private final Double lat;
-    private final Double lng;
+    private final BigDecimal lat;
+    private final BigDecimal lng;
 
 }
